BackGroundStage 연속처리

+++ Exam01/Exam01/Game_First_Stage.h
@@ -8,5 +8,5 @@ public:
 	void Update() override;
 	void Render() override;
 
-	float firstStageY = 0;
+	int firstStageY = 0;
 };
